Change to a working algorithm
The goal is to find the buy and sell time that will return the bigger
profit, but profit is calculated not on the difference between the sell
and buy price, but on the total money won minus invested.

In these cases we need to first calculate how much shares we can buy and
then sell them for profit. The total money won is calculated on the
following formula:

ROI = (investment / buy price) * sell price / investment

which is equivelent of (sell price / buy price) where ROI below 1 are
not profitable.